requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
paho-mqtt==1.6.1
numpy==1.24.3
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
cryptography==41.0.7
prometheus-fastapi-instrumentator==6.0.0
opencv-python==4.8.1.78
motor==3.3.1
azure-iot-device==2.15.0
azure-storage-blob==12.19.0
sentry-sdk==1.40.0